requirements.txt
asgiref==3.2.7
cachetools==4.1.1
certifi==2020.4.5.2
cffi==1.14.3
chardet==3.0.4
Django==3.0.7
django-cors-headers==3.3.0
django-settings-export==1.2.1
django-storages==1.10.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.2
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
idna==2.9
myql==1.2.7
oauthlib==3.1.0
Pillow==7.1.2
protobuf==3.13.0
py==1.8.1
pyaml==20.4.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyJWT==1.7.1
PyMySQL==0.9.3
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
rauth==0.7.3
requests==2.23.0
requests-oauthlib==1.3.0
rsa==4.6
six==1.15.0
sqlparse==0.3.1
ua-parser==0.10.0
urllib3==1.25.9
user-agents==2.1
xlwt==1.3.0
yahoo-oauth==1.0
